+++ src/components/calendar/elements/weekDays.tsx
@@ -1,0 +1,13 @@
+export default function WeekDays() {
+  return (
+    <div className="weekdays">
+      <div>Pon</div>
+      <div>Wt</div>
+      <div>Śr</div>
+      <div>Czw</div>
+      <div>Pt</div>
+      <div>Sob</div>
+      <div>Nie</div>
+    </div>
+  );
+}
