+++ src/components/graph/graph.tsx
@@ -4,7 +4,8 @@ import InfoWindow from "./infoWindow";
 import Calendar from "../calendar/calendar";
 import NursesWindow from "./nursesWindow";
 import ActualNurse from "./actualNurse";
-import ButtonProvider from "./buttonContext";
+import ButtonProvider from "../calendar/buttonContext";
+// import MonthProvider from "../calendar/monthContext";
 import "./graph.css";
 
 function Graph() {
