--- src/components/help.tsx
@@ -1,3 +1,0 @@
-export default function Help() {
-  return null;
-}
