+++ src/components/help/help.css
@@ -231,107 +231,6 @@ body {
   border-radius: 0.5rem;
 }
 /* 
-.tank {
-  font-family: "Poppins", sans-serif;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.buttons {
-  display: flex;
-  flex-direction: column;
-}
-.btn {
-  padding: 20px 50px;
-  font-size: 1.5rem;
-  position: relative;
-  margin: 32px;
-}
-.btn-1 {
-  background: black;
-  color: white;
-  border-radius: 30px;
-  transition: transform 0.3s ease;
-}
-.btn-1::after,
-.btn-1::before {
-  content: "";
-  position: absolute;
-  opacity: 0.3;
-  background: black;
-  border-radius: inherit;
-  width: 100%;
-  height: 100%;
-  left: 0;
-  bottom: 0;
-  z-index: -1;
-  transition: transform 0.3s ease;
-}
-.btn-1:hover {
-  transform: translate(12px, -12px);
-}
-.btn-1:hover::after {
-  transform: translate(-6px, 6px);
-}
-
-.btn-1:hover::before {
-  transform: translate(-12px, 12px);
-}
-
-.btn-2 {
-  color: black;
-}
-.btn-2::after,
-.btn-2::before {
-  border: 3px solid black;
-  content: "";
-  position: absolute;
-  width: calc(100%);
-  height: calc(100%);
-  left: 0;
-  bottom: 0;
-  z-index: -1;
-  transition: transform 0.3s ease;
-}
-.btn-2:hover::after {
-  transform: translate(-5px, -5px);
-}
-.btn-2:hover::before {
-  transform: translate(5px, 5px);
-}
-.btn-3 {
-  overflow: hidden;
-  color: white;
-  border-radius: 30px;
-  box-shadow: 0 0 0 0 rgba(143, 64, 248, 0.5), 0 0 0 0 rgba(39, 200, 255, 0.5);
-  transition: transform 0.3 ease, box-shadow 0.3s ease;
-}
-.btn-3::after {
-  content: "";
-  width: 400px;
-  height: 400px;
-  position: absolute;
-  top: -50px;
-  left: -100px;
-  background-color: #ff3cac;
-  background-image: linear-gradient(
-    225deg,
-    #27d86c 0%,
-    #26caf8 50%,
-    #c625d0 100%
-  );
-  z-index: -1;
-  transition: transform 0.5s ease;
-}
-.btn-3:hover {
-  transform: translate(0, -6px);
-  box-shadow: 10px -10px 25px 0 rgba(143, 64, 248, 0.5),
-    -10px 10px 25px 0 rgba(39, 200, 255, 0.5);
-}
-.btn-3:hover::after {
-  transform: rotate(180deg);
-} */
 
 .father {
   background-color: lightyellow;
