+++ src/components/help/help.tsx
@@ -1,0 +1,5 @@
+
+import PrevDays from '../calendar/elements/prevDays';
+export default function Help() {
+  return <PrevDays />;
+}
