+++ src/components/notFound.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 
 const NotFound = () => {
   return <h1>NOT FOUND</h1>;
