+++ src/components/nursesList/tableHeader.tsx
@@ -1,0 +1,16 @@
+import "./list.css";
+
+export function TableHeader() {
+  return (
+    <thead>
+      <tr>
+        <th scope="col">Imie i Nazwisko</th>
+        <th scope="col">Specka</th>
+        <th scope="col">Dni</th>
+        <th scope="col">Noce</th>
+        <th />
+        <th />
+      </tr>
+    </thead>
+  );
+}
